requirements.txt
opencv-python>=4.3.0.36
Pillow==7.0.0
progress==1.5
